+++ CHANGELOG.md
@@ -8,11 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Автоматическая генерация типов с помощью drf-spectacular + openapi-typescript
 
+## [0.1.0] - 2025-06-14
+### Added
 - Создана базовая модель Todo для старта 
 - В проект в проект Django установлен DRF c маршрутизацией через DefaultRouter
-
 ### Changed
-
 - Параметры проекта вынесены в переменные окружения и доступ к ним осуществляется через библиотеку через django-environ
 
+## [0.2.0] - 2025-06-14
+### Added
+- Добавлен frontend(react). Функционал простой отрисовка списка задач